+++ StreamVideoTests/Utilities/Dummy/EgressResponse+Dummy.swift
@@ -8,9 +8,15 @@ import StreamVideo
 extension EgressResponse {
     static func dummy(
         broadcasting: Bool = false,
+        frameRecording: FrameRecordingResponse? = nil,
         hls: EgressHLSResponse? = nil,
         rtmps: [EgressRTMPResponse] = []
     ) -> EgressResponse {
-        .init(broadcasting: broadcasting, hls: hls, rtmps: rtmps)
+        .init(
+            broadcasting: broadcasting,
+            frameRecording: frameRecording,
+            hls: hls,
+            rtmps: rtmps
+        )
     }
 }